- use int64 in issue numbers

+++ src/main/java/com/spotify/github/v3/Milestone.java
@@ -55,7 +55,7 @@ public interface Milestone extends CloseTracking {
 
   /** Number */
   @Nullable
-  Integer number();
+  Long number();
 
   /** The state of the milestone. Either open, closed, or all. */
   @Nullable
