+++ src/main/java/com/spotify/github/v3/issues/Issue.java
@@ -68,7 +68,7 @@ public interface Issue extends CloseTracking {
 
   /** Number. */
   @Nullable
-  Integer number();
+  Long number();
 
   /** Indicates the state of the issues to return. Can be either open, closed, or all. */
   @Nullable
